requirements.txt
alembic==1.14.1
aniso8601==10.0.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==8.1.8
colorama==0.4.4
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
duplicity==0.8.20
entrypoints==0.4
fasteners==0.14.1
filelock==3.16.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
httplib2==0.20.2
idna==2.8
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2024.4.0
platformdirs==4.3.6
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-docx==1.1.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.27.1
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
gunicorn
Flask-CORS
python-dotenv
Flask-JWT-Extended
Flask-Bcrypt

